+++ packages/cli/src/zosfiles/compare/Compare.definition.ts
@@ -11,6 +11,8 @@
 
 import {ICommandDefinition} from "@zowe/imperative";
 import {DatasetDefinition} from "./ds/Dataset.definition";
+import { LocalfileDatasetDefinition } from "./lfds/LocalfileDataset.definition";
+import {UssFileDefinition} from './uss/UssFile.definition';
 
 import i18nTypings from "../-strings-/en";
 
@@ -28,6 +30,8 @@ export const CompareDefinition: ICommandDefinition = {
     summary: strings.SUMMARY,
     description: strings.DESCRIPTION,
     children: [
-        DatasetDefinition
+        DatasetDefinition,
+        LocalfileDatasetDefinition,
+        UssFileDefinition
     ],
 };