+++ packages/cli/src/zosfiles/compare/uss/UssFile.definition.ts
@@ -17,30 +17,30 @@ import i18nTypings from "../../-strings-/en";
 const strings = (require("../../-strings-/en").default as typeof i18nTypings).COMPARE;
 
 /**
- * Compare data sets command definition containing its description, examples and/or options
+ * Compare uss-files command definition containing its description, examples and/or options
  * @type {ICommandDefinition}
  */
-export const DatasetDefinition: ICommandDefinition = {
-    name: "data-set",
-    aliases: ["ds"],
-    summary: strings.ACTIONS.DATA_SET.SUMMARY,
-    description: strings.ACTIONS.DATA_SET.DESCRIPTION,
+export const UssFileDefinition: ICommandDefinition = {
+    name: "uss-files",
+    aliases: ["uss"],
+    summary: strings.ACTIONS.USS_FILE.SUMMARY,
+    description: strings.ACTIONS.USS_FILE.DESCRIPTION,
     type: "command",
-    handler: __dirname + "/Dataset.handler",
+    handler: __dirname + "/UssFile.handler",
     profile: {
         optional: ["zosmf"],
     },
     positionals: [
         {
-            name: "dataSetName1",
-            description: strings.ACTIONS.DATA_SET.POSITIONALS.DATASETNAME1,
+            name: "ussFilePath1",
+            description: strings.ACTIONS.USS_FILE.POSITIONALS.USSFILEPATH1,
             type: "string",
             required: true
         },
         {
-            name: "dataSetName2",
+            name: "ussFilePath2",
             type: "string",
-            description: strings.ACTIONS.DATA_SET.POSITIONALS.DATASETNAME2,
+            description: strings.ACTIONS.USS_FILE.POSITIONALS.USSFILEPATH2,
             required: true
         }
     ],
@@ -49,22 +49,18 @@ export const DatasetDefinition: ICommandDefinition = {
         CompareOptions.binary2,
         CompareOptions.encoding,
         CompareOptions.encoding2,
-        CompareOptions.record,
-        CompareOptions.record2,
-        CompareOptions.volume,
-        CompareOptions.volume2,
         CompareOptions.seqnum,
         CompareOptions.contextLines,
         CompareOptions.browserView
     ],
     examples: [
         {
-            description: strings.ACTIONS.DATA_SET.EXAMPLES.EX1,
-            options: `"sys1.samplib(antptso)" "sys1.samplib(antxtso)"`
+            description: strings.ACTIONS.USS_FILE.EXAMPLES.EX1,
+            options: `"/u/**/**/*.ext" "/u/**/**/*.ext"`
         },
         {
-            description: strings.ACTIONS.DATA_SET.EXAMPLES.EX2,
-            options: `"sys1.samplib(antptso)" "sys1.samplib(antxtso)" --no-seqnum`
+            description: strings.ACTIONS.USS_FILE.EXAMPLES.EX2,
+            options: `"/u/**/**/*.ext" "/u/**/**/*.ext" --no-seqnum`
         }
     ]
 };